% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountSig.R
\name{CountSig}
\alias{CountSig}
\title{Determine if the peak is significant}
\usage{
CountSig(A, A.unc, B, n, m, k.alpha = 1.645)
}
\arguments{
\item{A}{net peak counts}

\item{A.unc}{net peak counts}

\item{B}{continuum counts}

\item{n}{peak width in channels}

\item{m}{channels used for background determination}

\item{k.alpha}{factor for probability interval, default is 1.645 for alpha of 0.05}
}
