% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalLimit.R
\name{CriticalLimit}
\alias{CriticalLimit}
\title{Determine the critical limit}
\usage{
CriticalLimit(B, n, m, k.alpha = 1.645)
}
\arguments{
\item{B}{continuum counts}

\item{n}{peak width in channels}

\item{m}{channels used for background determination}

\item{k.alpha}{factor for probability interval, default is 1.645 for alpha of 0.05}
}
\description{
"Is the net count significant" Gilmore (2008) Eq. 5.56 p.116
}
