% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMine.R
\name{DataMine}
\alias{DataMine}
\title{Read in Genie report files and calculate activities}
\usage{
DataMine(type = c("half jar", "full jar", "regression"), folder = NULL,
  csv = TRUE, raw = FALSE, filename = "DEFAULT")
}
\arguments{
\item{type}{type of efficiency to use}

\item{folder}{name of folder where reports are located. If NULL (default) user prompted for location}

\item{csv}{logical; write a csv of the output to the working directory}

\item{raw}{logical; returns raw data}

\item{filename}{filename of output csv. Defaults to folder name.}
}
\description{
Read in Genie report files and calculate activities
}

