% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorrange.R
\name{errorrange}
\alias{errorrange}
\title{Get range for vector with errors}
\usage{
errorrange(x, x.unc = 0.75, na.rm = T)
}
\arguments{
\item{x}{numeric vector}

\item{x.unc}{numeric vector of uncertainties same length as x}

\item{na.rm}{remove na? default TRUE}
}
\description{
Get range for vector with errors
}

