% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myaxis.R
\name{myaxis}
\alias{myaxis}
\title{Plot a custom axis for printing}
\usage{
myaxis(side, label.line = 0.75, append.label = "", col = "black")
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be drawn on. The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{label.line}{a numeric value specifying margin line to place tick labels}

\item{append.label}{a character string with which to append the axis tick label, ex. "\%"}
}
\description{
Plot a custom axis for printing
}

