% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsPb210.R
\name{xsPb210}
\alias{xsPb210}
\title{Calculate Excess Pb-210}
\usage{
xsPb210(PbAct = NA, PbActUnc = NA, SupPbAct = NA, SupPbActUnc = NA)
}
\arguments{
\item{PbAct}{vector of total Pb-210 activity values}

\item{PbActUnc}{vector of total Pb-210 activity error. Note: absolute not percent}

\item{SupPbAct}{vector or single value to be used for supported Pb-210}

\item{SupPbActUnc}{vector or single value of supported Pb-210 activity error. Note: absolute not percent}
}
\value{
Returns a two column vector of numerics with names xs210Pb and xs210Pb.uncer. NAs are placed where data is missing. If show.zero is true (default), then zeros are printed where xsPb210 is determined to be zero.
}
\description{
Calculate Excess Pb-210
}

