% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorPoints.R
\name{ErrorPoints}
\alias{ErrorPoints}
\title{Plot Horizontal and Vertical Line Segments on Existing Plot}
\usage{
ErrorPoints(x = NA, x.unc = NA, y = NA, y.unc = NA, show.hor = T,
  show.ver = T, ...)
}
\arguments{
\item{x}{a vector of x coordinates used as a centroid from which to draw segments.}

\item{x.unc}{a vector of uncertainites added and subracted from x to create segments.}

\item{y}{a vector of y coordinatesused as a centroid from which to draw segments.}

\item{y.unc}{a vector of uncertainites added and subracted from y to create segments.}

\item{show.hor}{logical; if TRUE, function draws horizontal segments}

\item{show.ver}{logical; if TRUE, function draws vertical segments}

\item{...}{further arguments passed to segments}
}
\description{
Plot Horizontal and Vertical Line Segments on Existing Plot
}

